chore: Move legacy footer and legal sections to its own component
